+++ src/main/java/com/tsp/new_tsp_admin/api/domain/portfolio/AdminPortFolioDTO.java
@@ -19,8 +19,8 @@ import java.util.List;
 @NoArgsConstructor
 @ApiModel
 public class AdminPortFolioDTO extends NewCommonDTO {
-    @ApiModelProperty(required = true, value = "rnum", hidden = true, example = "1")
-    private Integer rnum;
+    @ApiModelProperty(required = true, value = "rowNum", hidden = true, example = "1")
+    private Integer rowNum;
 
     @ApiModelProperty(required = true, value = "idx", hidden = true, example = "1")
     private Long idx;
